requirements.txt
# Core (audio decode uses TorchCodec; system FFmpeg required: apt install ffmpeg)
torch>=2.0.0
torchaudio>=2.0.0
# Hugging Face
huggingface_hub>=0.19.0
transformers>=4.36.0
datasets>=2.14.0
accelerate>=0.25.0
torchcodec
# PEFT (LoRA)
peft>=0.7.0

# Optional: for Trainer / training
# accelerate is needed by transformers Trainer
wandb
python-dotenv
safetensors